+++ lib/AL_USDMaya/AL/usdmaya/PluginRegister.h
@@ -25,9 +25,11 @@
 #include "AL/usdmaya/cmds/DebugCommands.h"
 #include "AL/usdmaya/cmds/EventCommand.h"
 #include "AL/usdmaya/cmds/LayerCommands.h"
+#include "AL/usdmaya/cmds/ListTranslators.h"
 #include "AL/usdmaya/cmds/ProxyShapeCommands.h"
 #include "AL/usdmaya/cmds/ProxyShapeSelectCommands.h"
 #include "AL/usdmaya/cmds/RendererCommands.h"
+#include "AL/usdmaya/cmds/SyncFileIOGui.h"
 #include "AL/usdmaya/cmds/UnloadPrim.h"
 #include "AL/usdmaya/fileio/Export.h"
 #include "AL/usdmaya/fileio/ExportTranslator.h"
@@ -43,6 +45,9 @@
 #include "AL/usdmaya/nodes/RendererManager.h"
 #include "AL/usdmaya/nodes/Transform.h"
 #include "AL/usdmaya/nodes/TransformationMatrix.h"
+
+#include "pxr/base/plug/plugin.h"
+#include "pxr/base/plug/registry.h"
 
 #include "maya/MDrawRegistry.h"
 #include "maya/MGlobal.h"
@@ -223,6 +228,7 @@ MStatus registerPlugin(AFnPlugin& plugin)
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::UsdDebugCommand);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::ListEvents);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::ListCallbacks);
+  AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::ListTranslators);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::Callback);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::TriggerEvent);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::DeleteCallbacks);
@@ -232,6 +238,7 @@ MStatus registerPlugin(AFnPlugin& plugin)
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::EventLookup);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::TranslatePrim);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::LayerManager);
+  AL_REGISTER_COMMAND(plugin, AL::usdmaya::cmds::SyncFileIOGui);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::fileio::ImportCommand);
   AL_REGISTER_COMMAND(plugin, AL::usdmaya::fileio::ExportCommand);
   AL_REGISTER_TRANSLATOR(plugin, AL::usdmaya::fileio::ImportTranslator);
@@ -273,6 +280,15 @@ MStatus registerPlugin(AFnPlugin& plugin)
   AL::maya::utils::MenuBuilder::addEntry("USD/Animated Geometry/Connect selected meshes to USD (animated)", "AL_usdmaya_meshAnimImport");
   CHECK_MSTATUS(AL::maya::utils::MenuBuilder::generatePluginUI(plugin, "AL_usdmaya"));
   AL::usdmaya::Global::onPluginLoad();
+
+  // Force all plugins to be loaded at startup time. Unless we load plugins upfront
+  // options will not be registered until the start of import or export, and won't be available in the GUI
+  PlugPluginPtrVector plugins = PlugRegistry::GetInstance().GetAllPlugins();
+  for(auto& plugin : plugins)
+  {
+    if(!plugin->IsLoaded())
+      plugin->Load();
+  }
   return status;
 }
 
@@ -306,6 +322,7 @@ MStatus unregisterPlugin(AFnPlugin& plugin)
     }
   }
 
+  AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::SyncFileIOGui);
   AL_UNREGISTER_COMMAND(plugin, AL::maya::utils::CommandGuiListGen);
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::InternalProxyShapeSelect);
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::ProxyShapePostSelect);
@@ -328,6 +345,7 @@ MStatus unregisterPlugin(AFnPlugin& plugin)
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::Callback);
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::ListCallbacks);
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::ListEvents);
+  AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::ListTranslators);
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::TriggerEvent);
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::DeleteCallbacks);
   AL_UNREGISTER_COMMAND(plugin, AL::usdmaya::cmds::CallbackQuery);