+++ lib/AL_USDMaya/AL/usdmaya/fileio/AnimationTranslator.cpp
@@ -29,6 +29,8 @@
 #include "maya/MFnMesh.h"
 #include "maya/MAnimUtil.h"
 #include "maya/MNodeClass.h"
+#include "maya/MMatrix.h"
+
 
 namespace AL {
 namespace usdmaya {
@@ -249,7 +251,9 @@ bool AnimationTranslator::isAnimatedTransform(const MObject& transformNode)
   while(currPath.pop() == MStatus::kSuccess && currPath.hasFn(MFn::kTransform) && inheritTransform(currPath))
   {
     if(areTransformAttributesConnected(currPath))
+    {
       return true;
+    }
   }
 
   return false;
@@ -266,10 +270,14 @@ void AnimationTranslator::exportAnimation(const ExporterParams& params)
   auto const endTransformAttrib = m_animatedTransformPlugs.end();
   auto const startMesh = m_animatedMeshes.begin();
   auto const endMesh = m_animatedMeshes.end();
+  auto const startWSM = m_worldSpaceOutputs.begin();
+  auto const endWSM = m_worldSpaceOutputs.end();
+
   if((startAttrib != endAttrib) ||
      (startAttribScaled != endAttribScaled) ||
      (startTransformAttrib != endTransformAttrib) ||
      (startMesh != endMesh) ||
+     (startWSM != endWSM) ||
      (!m_animatedNodes.empty()))
   {
     double increment = 1.0 / std::max(1U, params.m_subSamples);
@@ -306,6 +314,11 @@ void AnimationTranslator::exportAnimation(const ExporterParams& params)
       for(auto nodeAnim : m_animatedNodes)
       {
         nodeAnim.m_translator->exportCustomAnim(nodeAnim.m_path, nodeAnim.m_prim, timeCode);
+      }
+      for(auto it = startWSM; it != endWSM; ++it)
+      {
+        MMatrix mat = it->first.inclusiveMatrix();
+        it->second.Set(*(const GfMatrix4d*)&mat, timeCode);
       }
     }
   }